+++ src/plonemeeting/portal/core/migrations/migrate_to_1004.py
@@ -1,13 +1,30 @@
 # -*- coding: utf-8 -*-
-import logging
 from copy import deepcopy
-
 from imio.migrator.migrator import Migrator
+from plonemeeting.portal.core.utils import set_constrain_types
+
+import logging
+
 
 logger = logging.getLogger("plonemeeting.portal.core")
 
 
 class MigrateTo1004(Migrator):
+
+    def _apply_folder_constraints(self):
+        """
+        Apply contraints on Folders.
+        """
+        logger.info("Apply contraints on Faceted Folders")
+        brains = self.catalog(portal_type="Institution")
+        for brain in brains:
+            institution = brain.getObject()
+            for folder_id in ("seances", "decisions"):
+                folder = institution.get(folder_id)
+                if folder:
+                    set_constrain_types(folder, [])
+        logger.info("Done.")
+
     def _init_long_representatives_in_charge(self):
         """
         Initialize the new attribute long_representatives_in_charge
@@ -23,6 +40,7 @@ class MigrateTo1004(Migrator):
     def run(self):
         logger.info("Migrating to plonemeeting.portal 1004...")
         self._init_long_representatives_in_charge()
+        self._apply_folder_constraints()
         self.refreshDatabase(catalogs=False,
                              workflows=True,
                              workflowsToUpdate=["institution_workflow"])
@@ -31,7 +49,9 @@ class MigrateTo1004(Migrator):
 def migrate(context):
     """
     This migration function will:
-       1) Initialize the new attribute long_representatives_in_charge ;
+       1) Initialize the new attribute long_representatives_in_charge;
+       2) Apply contraints on Folders;
+       3) Update security for elements using institution_workflow.
     """
     migrator = MigrateTo1004(context)
     migrator.run()